requirements.txt
beautifulsoup4==4.11.1
Flask==2.2.2
pandas==1.5.0
pinecone-client==2.0.13
requests==2.28.1
sentence-transformers==2.2.2
tqdm==4.64.1